backend/voiceTranslation/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
transformers
torch
torchaudio
soundfile
librosa
